Dynamically Allocate and Free window_class using Marshal
I wasn't able to reproduce this on Windows, but on Linux, there is an issue where the data changes for no logical reason. I think it has to do with how the garbage collector works in csharp.

As a consequence of this change, I moved the implementation of the IDisposable interface to the DockSpaceWindow instead.